LayoutTests/platform/ios-wk2/fast/forms/validation-messages-expected.txt

Required text input:
Required checkbox input:
Required radio input:
Required radio input:
Required file input:
Required email input:
Required url input:
Required input with pattern:
Required input with minlength=100:
Required range with min=5:
Required range with max=5:
Required range with step=3 / min=0:
Tests the HTML form validation messages being shown on UI side.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS message is "Fill out this field"
PASS message is "Select this checkbox"
PASS message is "Select one of these options"
PASS message is "Select a file"
PASS message is "Enter an email address"
PASS message is "Enter a URL"
PASS message is "Match the requested format"
FAIL message should be Use at least 100 characters. Was Fill out this field.
PASS message is "Value must be greater than or equal to 5"
PASS message is "Value must be less than or equal to 5"
PASS message is "Enter a valid value"
PASS successfullyParsed is true
Some tests failed.

TEST COMPLETE
